requirements.txt
pytube==15.0.0
python-telegram-bot==20.3